pyproject.toml
[project]
name = "perudo-wasm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.3",
    "pygbag>=0.9.2",
]
